DeepLearning/Tensorflow/Model_Demo/requirements.txt
matplotlib
tensorflow_datasets
